refactor: cleaned up tests directory
